backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
beautifulsoup4==4.13.5
requests==2.32.5
aiohttp==3.12.15
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1
cloudscraper==1.2.71